packages/kubernetes/pyproject.toml
[project]
name = "pragmatiks-kubernetes-provider"
version = "0.1.0"
description = "Generic Kubernetes resources using lightkube"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pragmatiks-sdk>=0.6.0",
    "pragmatiks-gcp-provider>=0.55.0",
    "lightkube>=0.15.0",
    "google-auth>=2.30.0",
    "requests>=2.32.0",
]

authors = [
    { name = "Pragmatiks", email = "team@pragmatiks.io" },
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.0",
]

[tool.pragma]
# Provider metadata - used by the platform for discovery and deployment
provider = "kubernetes"
package = "kubernetes_provider"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_files = ["pyproject.toml:^version"]
tag_format = "kubernetes-v$version"
update_changelog_on_bump = true
changelog_file = "CHANGELOG.md"

[tool.ruff]
line-length = 120

[tool.uv.build-backend]
module-name = "kubernetes_provider"

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"
